backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-dotenv==1.0.1
google-genai==1.5.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.155.0
pyaudio==0.2.14
pydantic==2.10.0
numpy==2.1.0
